// contracts/generated/typescript/models/EntityFederationLink.ts
export interface EntityFederationLink {
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    wikiId: string;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    localEntityId: string;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    localCanonicalKey: string;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    localLabel: string;
    /**
     *
     * @type {EntityFederationLinkLocalStatusEnum}
     * @memberof EntityFederationLink
     */
    localStatus: EntityFederationLinkLocalStatusEnum;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    remoteWikiId: string;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    remoteWikiKey: string;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    remoteWikiName: string;
    /**
     *
     * @type {EntityFederationLinkRemoteWikiStatusEnum}
     * @memberof EntityFederationLink
     */
    remoteWikiStatus: EntityFederationLinkRemoteWikiStatusEnum;
    /**
     *
     * @type {EntityFederationLinkRemoteTrustLevelEnum}
     * @memberof EntityFederationLink
     */
    remoteTrustLevel: EntityFederationLinkRemoteTrustLevelEnum;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    remoteEntityId: string;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    remoteCanonicalKey: string;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    remoteLabel: string;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    remoteEntityUrl: string;
    /**
     *
     * @type {EntityFederationLinkRelationTypeEnum}
     * @memberof EntityFederationLink
     */
    relationType: EntityFederationLinkRelationTypeEnum;
    /**
     *
     * @type {EntityFederationLinkVerificationStatusEnum}
     * @memberof EntityFederationLink
     */
    verificationStatus: EntityFederationLinkVerificationStatusEnum;
    /**
     *
     * @type {EntityFederationLinkStatusEnum}
     * @memberof EntityFederationLink
     */
    status: EntityFederationLinkStatusEnum;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof EntityFederationLink
     */
    metadata: { [key: string]: any; };
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    createdBy: string;
    /**
     *
     * @type {string}
     * @memberof EntityFederationLink
     */
    updatedBy: string;
    /**
     *
     * @type {Date}
     * @memberof EntityFederationLink
     */
    createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof EntityFederationLink
     */
    updatedAt: Date;
}


/**
 * @export
 */
export const EntityFederationLinkLocalStatusEnum = {
    Active: 'active',
    Merged: 'merged',
    Deleted: 'deleted'
} as const;
export type EntityFederationLinkLocalStatusEnum = typeof EntityFederationLinkLocalStatusEnum[keyof typeof EntityFederationLinkLocalStatusEnum];

/**
 * @export
 */
export const EntityFederationLinkRemoteWikiStatusEnum = {
    Active: 'active',
    Disabled: 'disabled'
} as const;
export type EntityFederationLinkRemoteWikiStatusEnum = typeof EntityFederationLinkRemoteWikiStatusEnum[keyof typeof EntityFederationLinkRemoteWikiStatusEnum];

/**
 * @export
 */
export const EntityFederationLinkRemoteTrustLevelEnum = {
    Untrusted: 'untrusted',
    Reference: 'reference',
    Trusted: 'trusted'
} as const;
export type EntityFederationLinkRemoteTrustLevelEnum = typeof EntityFederationLinkRemoteTrustLevelEnum[keyof typeof EntityFederationLinkRemoteTrustLevelEnum];

/**
 * @export
 */
export const EntityFederationLinkRelationTypeEnum = {
    SameAs: 'same_as',
    Broader: 'broader',
    Narrower: 'narrower',
    Related: 'related'
} as const;
export type EntityFederationLinkRelationTypeEnum = typeof EntityFederationLinkRelationTypeEnum[keyof typeof EntityFederationLinkRelationTypeEnum];

/**
 * @export
 */
export const EntityFederationLinkVerificationStatusEnum = {
    Unverified: 'unverified',
    HumanVerified: 'human_verified',
    Disputed: 'disputed'
} as const;
export type EntityFederationLinkVerificationStatusEnum = typeof EntityFederationLinkVerificationStatusEnum[keyof typeof EntityFederationLinkVerificationStatusEnum];

/**
 * @export
 */
export const EntityFederationLinkStatusEnum = {
    Active: 'active',
    Deprecated: 'deprecated'
} as const;
export type EntityFederationLinkStatusEnum = typeof EntityFederationLinkStatusEnum[keyof typeof EntityFederationLinkStatusEnum];